Add OrdersController, Checkout, and MyOrders pages for order management
- Implemented OrdersController to handle order retrieval and placement.
- Created Checkout.razor for reviewing and placing orders.
- Developed MyOrders.razor to display user's order history with status.
- Integrated HTTP client for API communication and order state management.

+++ Model/Pizza.cs
@@ -26,7 +26,7 @@ namespace BlazingPizza.Model
 
         public decimal GetBasePrice()
         {
-            return ((decimal)Size / (decimal)DefaultSize) * Special.BasePrice;
+            return (decimal) Size / DefaultSize * Special.BasePrice;
         }
 
         public decimal GetTotalPrice()
